refactor: extract FAD component
