+++ src/App.js
@@ -1,67 +1,11 @@
-import {
-  Fragment,
-  createElement,
-  useState,
-  useContext,
-  useMemo,
-  useEffect,
-} from './lib/react.js';
-import { createPortal } from './lib/react-dom.js';
+import { createElement, useState, useEffect } from './lib/react.js';
 import { CONFIG_KEY, THEMES } from './config/constants.js';
 import { concernedCLIConfig } from './config/cli.js';
 import ThemeContext from './context/theme.js';
 import { useLegacyCleaner } from './hooks/utils/use-legacy-cleaner.js';
-import { useFADStatus } from './hooks/host/use-fad-status.js';
 import { useTurntablePlayState } from './hooks/features/use-turntable-play-state.js';
-import { useFADSideEffect } from './hooks/features/use-fad-side-effect.js';
-import TrackHeart from './components/host-aware/track-heart.js';
-import SongPreview from './components/host-aware/song-preview.js';
 import Main from './components/injected/main.js';
-
-const FAD = () => {
-  const { theme } = useContext(ThemeContext);
-  useFADSideEffect();
-
-  const containers = useMemo(() => {
-    const fad = document.querySelector('#full-app-display');
-    return {
-      fad,
-      fadFg: fad.querySelector('#fad-foreground'),
-    };
-  }, []);
-
-  return createElement(
-    Fragment,
-    null,
-    createPortal(
-      createElement(
-        Fragment,
-        null,
-        createElement(SongPreview, {
-          containerClassName: 'fad-song-preview',
-        }),
-        theme === THEMES.DARK &&
-          createElement('div', {
-            id: 'fad-mask',
-          })
-      ),
-      containers.fad
-    ),
-    createPortal(
-      createElement(TrackHeart, {
-        className: 'fad-track-heart',
-      }),
-      containers.fadFg
-    )
-  );
-};
-
-const FADPortals = () => {
-  const status = useFADStatus();
-
-  if (!status) return null;
-  return createElement(FAD);
-};
+import FAD from './components/injected/fad.js';
 
 const App = () => {
   const { exts } = concernedCLIConfig;
@@ -94,7 +38,7 @@ const App = () => {
       },
     },
     createElement(Main),
-    exts.fullAppDisplay && createElement(FADPortals)
+    exts.fullAppDisplay && createElement(FAD)
   );
 };
