+++ src/index.css
@@ -53,37 +53,6 @@
   .main-trackCreditsModal-closeBtn path {
     fill: currentcolor;
   }
-
-  .fad-track-heart {
-    color: #ccc;
-  }
-
-  .fad-song-preview button {
-    color: #ccc;
-
-    &:hover {
-      color: #fff;
-    }
-  }
-}
-
-#full-app-display {
-  background-color: #2c2c2c;
-}
-
-#fad-mask {
-  width: 100vw;
-  height: 100vh;
-  position: absolute;
-  top: 0;
-  left: 0;
-  background-color: rgb(9 9 9 / 50%);
-  z-index: -2;
-}
-
-#fad-body {
-  height: calc(100vh - 120px) !important;
-  margin: 60px 0;
 }
 
 [data-testid='track-visual-enhancement'] .cover-art-image,
@@ -160,118 +129,7 @@
   z-index: -1;
 }
 
-#fad-details {
-  max-width: 520px !important;
-}
-
-#fad-details #fad-title {
-  font-size: 30px;
-}
-
-#fad-details #fad-artist {
-  margin-top: 4px;
-  font-size: 18px;
-}
-
-#fad-details #fad-album,
-#fad-details #fad-release-date {
-  margin: 6px 16px 0 0;
-  display: inline-block;
-  color: #ccc;
-}
-
-#fad-details #fad-album {
-  font-size: 13px;
-}
-
-#fad-details #fad-release-date {
-  font-size: 11px;
-}
-
-#fad-details #fad-artist > *,
-#fad-details #fad-album > *,
-#fad-details #fad-release-date > *,
-#fad-details #fad-status > #fad-controls > * > svg {
-  vertical-align: middle;
-}
-
-#fad-details #fad-artist > svg {
-  width: 20px;
-  height: 20px;
-}
-
-#fad-details #fad-album > svg {
-  width: 15px;
-  height: 15px;
-  margin-left: 4px;
-  margin-right: 6px;
-}
-
-#fad-details #fad-release-date > svg {
-  width: 13px;
-  height: 13px;
-  margin-left: 4px;
-  margin-right: 6px;
-}
-
-#fad-play > svg {
-  width: 24px;
-  height: 24px;
-}
-
-#fad-controls > button > svg {
-  fill: #ccc;
-}
-
-#fad-controls > button:hover > svg {
-  fill: #fff;
-}
-
-#fad-progress-container {
-  font-size: 12px;
-}
-
-#fad-elapsed,
-#fad-duration {
-  min-width: 32px !important;
-}
-
-#fad-progress {
-  height: var(--custom-progress-bar-height) !important;
-  background-color: rgb(200 200 200 / 30%) !important;
-}
-
-#fad-progress-inner {
-  background-color: var(--spice-button) !important;
-  box-shadow: unset !important;
-}
-
-#fad-thumb {
-  width: var(--custom-progress-bar-slider-size) !important;
-  height: var(--custom-progress-bar-slider-size) !important;
-  top: calc(
-    0px -
-      (
-        var(--custom-progress-bar-slider-size) -
-          var(--custom-progress-bar-height)
-      ) /
-      2
-  ) !important;
-}
-
-/* song preview */
-.fad-song-preview {
-  width: 100%;
-  position: absolute;
-  bottom: 20px;
-}
-
 /* Lyrics Plus */
-#fad-lyrics-plus-container,
-.lyrics-lyricsContainer-LyricsContainer.fad-enabled {
-  height: inherit !important;
-}
-
 .lyrics-lyricsContainer-LyricsLine {
   font-weight: 600 !important;
 }
@@ -285,25 +143,6 @@
   background-color: rgb(6 6 6 / 20%);
   outline: 1px solid rgb(150 150 150 / 10%);
   box-shadow: 0 0 20px rgb(150 150 150 / 20%);
-}
-
-/* Responsive */
-@media (width <= 908px) {
-  #fad-foreground {
-    flex-wrap: wrap;
-    align-content: center;
-  }
-
-  #fad-details {
-    padding-top: 50px;
-  }
-}
-
-@media (width >= 1460px) and (height >= 960px) {
-  #fad-foreground,
-  .main-trackCreditsModal-container {
-    transform: scale(1.2) !important;
-  }
 }
 
 @keyframes rotate-turntable {
